Pedantic (promote simplex noise instead of perlin)
Most cases where I see Perlin noise used, appear to be more because the user heard about it first, and not because it was best choice for the application. Attempting to help the cause in a small way, by changing the issue template to reference simplex noise instead of perlin noise.