Add and use some new OpenGL extensions
This commit adds 2 new extensions to the GLAD loader, GL_KHR_debug_output
and GL_KHR_no_error.
GL_KHR_debug_output allows for seamless debugging of OpenGL code,
which means that in case of an issue it should become trivial to
identify the source of the problem.
GL_KHR_no_error is an extension which adds a new context creation flag.
This flag tells the application that it should not try and generate new glErrors
which should increase the performance slightly in release modes.